[DEV-3696] Rename
